requirements.txt
flask
werkzeug
PyPDF2
python-docx
pandas
openpyxl